+++ cdm/src/main/java/ucar/nc2/iosp/uf/Ray.java
@@ -32,15 +32,16 @@
  */
 package ucar.nc2.iosp.uf;
 
-import java.io.IOException;
-import java.nio.ByteBuffer;
-import java.util.Calendar;
-import java.util.Date;
-import java.util.HashMap;
-
-import ucar.ma2.IndexIterator;
-import ucar.ma2.Range;
 import ucar.unidata.io.RandomAccessFile;
+import ucar.ma2.Range;
+import ucar.ma2.IndexIterator;
+
+import java.nio.ByteBuffer;
+import java.util.Date;
+import java.util.Calendar;
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.io.IOException;
 
 /**
  * Created by IntelliJ IDEA.
@@ -493,9 +494,9 @@ public class Ray {
     public static int bytesToShort(byte a, byte b, boolean swapBytes) {
         // again, high order bit is expressed left into 32-bit form
         if (swapBytes) {
-            return (a & 0xff) + (b << 8);
+            return (a & 0xff) + ((int)b << 8);
         } else {
-            return (a << 8) + (b & 0xff);
+            return ((int)a << 8) + (b & 0xff);
         }
     }
     public static int bytesToInt(byte [] bytes, boolean swapBytes) {
